requirements.txt
# Core dependencies
scrapy>=2.11.0
selenium>=4.15.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pgvector>=0.4.0

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
typer>=0.9.0
rich>=13.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Web scraping enhancements
fake-useragent>=1.4.0
undetected-chromedriver>=3.5.5
webdriver-manager>=4.0.0
curl-cffi>=0.6.0  # TLS fingerprinting for bypass HTTP 403 (OccMundial)

# API and HTTP
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiohttp>=3.8.0
httpx>=0.24.0
redis>=5.0.0
flower>=2.0.0

# Text processing
nltk>=3.8.0
spacy>=3.6.0

# Machine Learning and embeddings
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
umap-learn>=0.5.5
hdbscan>=0.8.33  # Works on ARM64 (tested and verified)

# LLM inference
llama-cpp-python>=0.2.0  # CPU/GPU inference for GGUF models
# vllm>=0.3.0  # Optional: High-performance GPU inference (uncomment if using GPU)
openai>=1.0.0  # Optional: OpenAI API fallback

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Scheduling and automation
apscheduler>=3.10.0
celery>=5.3.0
psutil>=5.9.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Docker and deployment
docker>=6.1.0
kubernetes>=27.0.0

# Additional utilities
tqdm>=4.65.0
click>=8.1.0
colorama>=0.4.6

# Database enhancements
pgvector  # PostgreSQL vector extension support
